Moved example to use Hooks
I was going to recommend this tutorial to a friend who is learning React, but realized that it was using classes while I was recommending him to go with Hooks. React's official position is that [React Hooks are preferred over classes](https://reactjs.org/docs/hooks-faq.html#should-i-use-hooks-classes-or-a-mix-of-both), so this PR changes the documentation to use hooks instead of life-cycle methods.

Please let me know if I should continue with the rest of examples in this file.